# ShopAPI/ShopAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Shop API",
    "version": "v1"
  },
  "paths": {
    "/api/Account/Register": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/Login": {
      "post": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/ForgotPassword": {
      "post": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/ResetPassword": {
      "post": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/Profile": {
      "get": {
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Account/UpdateProfile": {
      "put": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserProfileUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Account/UpdatePassword": {
      "put": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "password",
            "required": true,
            "type": "string",
            "maxLength": 100,
            "minLength": 8,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/api/Contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ContactDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "definitions": {
    "ContactDto": {
      "required": [
        "email",
        "firstName",
        "lastName",
        "message"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "format": "email",
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phone": {
          "maxLength": 100,
          "type": "string"
        },
        "subjectId": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "maxLength": 4000,
          "minLength": 20,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UserDto": {
      "required": [
        "address",
        "email",
        "firstName",
        "lastName",
        "password"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "format": "email",
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phone": {
          "maxLength": 20,
          "type": "string"
        },
        "address": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "minLength": 8,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserProfileUpdateDto": {
      "required": [
        "address",
        "email",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "format": "email",
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phone": {
          "maxLength": 20,
          "type": "string"
        },
        "address": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter token"
    }
  }
}